Ignore dataclass fields with `init=False`
It turns out the current specialization and binding implementations
actually handle the `init` exclusion correctly, but I committed the
test cases for it anyway, to make sure this continues to work if
the implementation is ever changed.

The template formatter did need an additional check to exlude fields
with `init=False`.

Closes: #2