Drop `SLF` Ruff rule
There is a relatively high number of false positives, as Ruff does not
have type information and therefore can only detect the `self` literal
and not other objects of the same type.

This rule is also one that is hard to violate accidentally: if you are
using private API outside of the class, it's typically a conscious
choice.